requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
SQLAlchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose==3.3.0
insightface==0.7.3
onnxruntime==1.17.1
numpy==1.26.4
Pillow==10.3.0
